Basket API: Allow re-order for anonymous customers
Mainly for staff. This is needed for shop staff that needs
to re-order some past order for anonymous customer.
